+++ public/ui/objectExplorer.js
@@ -1,34 +1,23 @@
 import { component, html, css } from "../component.js"
-
-function createHtml(obj, indentLevel = 0) {
-  if (obj === null) {
-    return html`<b>null</b>`
-  }
-
-  if (obj === undefined) {
-    return html`<b>undefined</b>`
-  }
-
-  if (Array.isArray(obj)) {
-    return html`[${obj.map((o) => createHtml(o, indentLevel + 1))}]`
-  }
-
-  if (typeof obj === "object") {
-    return Object.keys(obj).map(
-      (k) =>
-        html`<div style="margin-left: ${indentLevel * 10}px">
-          ${k}: ${createHtml(obj[k], indentLevel + 1)}
-        </div>`
-    )
-  }
-
-  return html`<b style="margin-left: ${indentLevel * 10}px">${obj}</b>`
-}
+import formatSize from "../formatSize.js"
 
 component(
   `x-object-explorer`,
   await css(import.meta.resolve(`./objectExplorer.css`)),
-  function ObjectExplorer({ obj }) {
-    return html`${JSON.stringify(obj, null, 4)}`
+  function ObjectExplorer({ selectedTorrents }) {
+    if (!selectedTorrents.length) {
+      return html``
+    }
+
+    if (selectedTorrents.length === 1) {
+      return html`${JSON.stringify(selectedTorrents[0], null, 4)}`
+    }
+
+    const items = [
+      `Selected: ${selectedTorrents.length}`,
+      `Total size: ${formatSize(selectedTorrents.map((t) => t.totalSize).reduce((a, b) => a + b, 0))}`,
+    ]
+
+    return html`${items.map((item) => `  ${item}`).join(`\n`)}`
   }
 )