+++ public/ui/torrents/Selections.js
@@ -1,13 +1,12 @@
 import { useMemo, useState } from "../../component.js"
 
-export default function selectionManager({ torrents, setSelectedId }) {
-  const [selections, setSelections] = useState([])
-
+export default function selectionManager({
+  torrents,
+  selections,
+  setSelections,
+}) {
   function onClickRow(e, _id) {
     const id = _id || +this.dataset.id
-
-    // just to notify x-main so that it can show details if necessary
-    setSelectedId(id)
 
     if (!e.shiftKey && !e.ctrlKey) {
       setSelections([id])